Add multiple pomo timers
