+++ css/main.css
@@ -4,7 +4,7 @@
   margin-bottom: 40px;
 }
 
-.pomo {
+.pomo__clock {
   display: flex;
   justify-content: center;
   align-items: center;
