+++ js/util.js
@@ -1,0 +1,3 @@
+export const addZero = (num) => {
+  return `0${num}`.slice(-2);
+};
